+++ src/SuperSimpleTcp/SimpleTcpClientSettings.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Net.Security;
 
 namespace SuperSimpleTcp
 {
@@ -119,6 +120,22 @@ namespace SuperSimpleTcp
         /// </summary>
         public bool MutuallyAuthenticate = true;
 
+        /// <summary>
+        /// Enable or disable whether the data receiver thread fires the DataReceived event from a background task.
+        /// The default is enabled.
+        /// </summary>
+        public bool UseAsyncDataReceivedEvents = true;
+
+        /// <summary>
+        /// Enable or disable checking certificate revocation list during the validation process.
+        /// </summary>
+        public bool CheckCertificateRevocation = true;
+
+        /// <summary>
+        /// Delegate responsible for validating a certificate supplied by a remote party.
+        /// </summary>
+        public RemoteCertificateValidationCallback CertificateValidationCallback = null;
+
         #endregion
 
         #region Private-Members
@@ -141,6 +158,14 @@ namespace SuperSimpleTcp
         {
 
         }
+
+        #endregion
+
+        #region Public-Methods
+
+        #endregion
+
+        #region Private-Methods
 
         #endregion
     }