+++ src/SuperSimpleTcp/SimpleTcpServer.cs
@@ -749,12 +749,19 @@ namespace SuperSimpleTcp
                         {
                             client.SslStream = new SslStream(client.NetworkStream, false, new RemoteCertificateValidationCallback(AcceptCertificate));
                         }
+                        else if(_settings.CertificateValidationCallback != null)
+                        {
+                            client.SslStream = new SslStream(client.NetworkStream, false, new RemoteCertificateValidationCallback(_settings.CertificateValidationCallback));
+                        }
                         else
                         {
                             client.SslStream = new SslStream(client.NetworkStream, false);
                         }
 
-                        bool success = await StartTls(client).ConfigureAwait(false);
+                        CancellationTokenSource tlsCts = CancellationTokenSource.CreateLinkedTokenSource(_listenerToken, _token);
+                        tlsCts.CancelAfter(3000);
+
+                        bool success = await StartTls(client, tlsCts.Token).ConfigureAwait(false);
                         if (!success)
                         {
                             client.Dispose();
@@ -807,7 +814,7 @@ namespace SuperSimpleTcp
             _isListening = false;
         }
 
-        private async Task<bool> StartTls(ClientMetadata client)
+        private async Task<bool> StartTls(ClientMetadata client, CancellationToken token)
         {
             try
             {
@@ -815,7 +822,7 @@ namespace SuperSimpleTcp
                     _sslCertificate,
                     _settings.MutuallyAuthenticate,
                     SslProtocols.Tls12,
-                    !_settings.AcceptInvalidCertificates).ConfigureAwait(false);
+                    _settings.CheckCertificateRevocation).ConfigureAwait(false);
 
                 if (!client.SslStream.IsEncrypted)
                 {
@@ -840,7 +847,15 @@ namespace SuperSimpleTcp
             }
             catch (Exception e)
             {
-                Logger?.Invoke($"{_header}client {client.IpPort} SSL/TLS exception: {Environment.NewLine}{e}");
+                if (e is TaskCanceledException || e is OperationCanceledException)
+                {
+                    Logger?.Invoke($"{_header}client {client.IpPort} timeout during SSL/TLS establishment");
+                }
+                else
+                {
+                    Logger?.Invoke($"{_header}client {client.IpPort} SSL/TLS exception: {Environment.NewLine}{e}");
+                }
+
                 client.Dispose();
                 return false;
             }
@@ -884,7 +899,16 @@ namespace SuperSimpleTcp
                         continue;
                     }
 
-                    _ = Task.Run(() => _events.HandleDataReceived(this, new DataReceivedEventArgs(ipPort, data)), linkedCts.Token);
+                    Action action = () => _events.HandleDataReceived(this, new DataReceivedEventArgs(ipPort, data));
+                    if (_settings.UseAsyncDataReceivedEvents)
+                    {
+                        _ = Task.Run(action, linkedCts.Token);
+                    }
+                    else
+                    {
+                        action.Invoke();
+                    }
+
                     _statistics.ReceivedBytes += data.Count;
                     UpdateClientLastSeen(client.IpPort);
                 }